third_party/WebKit/LayoutTests/animations/prefixed/keyframes-cssom-prefixed-02-expected.txt

Test the CSSOM of @-webkit-keyframes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS keyframeRule.toString() is "[object CSSKeyframesRule]"
PASS keyframeRule.type is 7
PASS keyframeRule.name is "anim"
PASS keyframeRule.findRule('from').cssText is "0% { left: 200px; }"
PASS keyframeRule.findRule('to').cssText is "100% { left: 300px; }"
PASS keyframeRule.findRule('from') is null
PASS keyframeRule.findRule('from').cssText is "0% { left: 200px; }"
PASS keyframeRule.cssRules[0].toString() is "[object CSSKeyframeRule]"
PASS keyframeRule.cssRules[0].cssText is "100% { left: 300px; }"
PASS keyframeRule.cssRules[0].keyText is "100%"
PASS keyframeRule.cssRules[0].style.cssText is "left: 300px;"
PASS keyframeRule.cssRules[1].toString() is "[object CSSKeyframeRule]"
PASS keyframeRule.cssRules[1].cssText is "0% { left: 200px; }"
PASS keyframeRule.cssRules[1].keyText is "0%"
PASS keyframeRule.cssRules[1].style.cssText is "left: 200px;"
PASS new CSSKeyframesRule().name threw exception TypeError: Illegal constructor.
PASS successfullyParsed is true

TEST COMPLETE
